pyproject.toml
[tool.poetry]
name = "utterexhaustedrobodoc"
version = "1.0.0"
description = "A Simple discord bot about player duels and some other stuff too"
authors = ["@engineer tf3#0776"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
